+++ src/engine/aggregator.go
@@ -21,7 +21,10 @@ type Aggregator interface {
 	ColumnNames() []string
 }
 
-type AggregatorInitializer func(*parser.SelectQuery, *parser.Value) (Aggregator, error)
+// Initialize a new aggregator given the query, the function call of
+// the aggregator and the default value that should be returned if
+// the bucket doesn't have any points
+type AggregatorInitializer func(*parser.SelectQuery, *parser.Value, *parser.Value) (Aggregator, error)
 
 var registeredAggregators = make(map[string]AggregatorInitializer)
 
@@ -99,7 +102,8 @@ type StandardDeviationRunning struct {
 
 type StandardDeviationAggregator struct {
 	AbstractAggregator
-	running map[string]map[interface{}]*StandardDeviationRunning
+	running      map[string]map[interface{}]*StandardDeviationRunning
+	defaultValue *protocol.FieldValue
 }
 
 func (self *StandardDeviationAggregator) AggregatePoint(series string, group interface{}, p *protocol.Point) error {
@@ -142,6 +146,12 @@ func (self *StandardDeviationAggregator) ColumnNames() []string {
 
 func (self *StandardDeviationAggregator) GetValues(series string, group interface{}) [][]*protocol.FieldValue {
 	r := self.running[series][group]
+	if r.count == 0 {
+		return [][]*protocol.FieldValue{
+			[]*protocol.FieldValue{self.defaultValue},
+		}
+	}
+
 	eX := r.totalX / float64(r.count)
 	eX *= eX
 	eX2 := r.totalX2 / float64(r.count)
@@ -153,7 +163,22 @@ func (self *StandardDeviationAggregator) GetValues(series string, group interfac
 	}
 }
 
-func NewStandardDeviationAggregator(q *parser.SelectQuery, v *parser.Value) (Aggregator, error) {
+func wrapDefaultValue(defaultValue *parser.Value) (*protocol.FieldValue, error) {
+	if defaultValue == nil {
+		return nil, nil
+	}
+
+	switch defaultValue.Type {
+	case parser.ValueInt:
+		v, _ := strconv.Atoi(defaultValue.Name)
+		value := int64(v)
+		return &protocol.FieldValue{Int64Value: &value}, nil
+	default:
+		return nil, fmt.Errorf("Unknown type %s", defaultValue.Type)
+	}
+}
+
+func NewStandardDeviationAggregator(q *parser.SelectQuery, v *parser.Value, defaultValue *parser.Value) (Aggregator, error) {
 	if len(v.Elems) != 1 {
 		return nil, common.NewQueryError(common.WrongNumberOfArguments, "function stddev() requires exactly one argument")
 	}
@@ -162,11 +187,16 @@ func NewStandardDeviationAggregator(q *parser.SelectQuery, v *parser.Value) (Agg
 		return nil, common.NewQueryError(common.InvalidArgument, "function stddev() doesn't work with wildcards")
 	}
 
+	value, err := wrapDefaultValue(defaultValue)
+	if err != nil {
+		return nil, err
+	}
 	return &StandardDeviationAggregator{
 		AbstractAggregator: AbstractAggregator{
 			value: v.Elems[0],
 		},
-		running: make(map[string]map[interface{}]*StandardDeviationRunning),
+		running:      make(map[string]map[interface{}]*StandardDeviationRunning),
+		defaultValue: value,
 	}, nil
 }
 
@@ -176,8 +206,9 @@ func NewStandardDeviationAggregator(q *parser.SelectQuery, v *parser.Value) (Agg
 
 type DerivativeAggregator struct {
 	AbstractAggregator
-	firstValues map[string]map[interface{}]*protocol.Point
-	lastValues  map[string]map[interface{}]*protocol.Point
+	firstValues  map[string]map[interface{}]*protocol.Point
+	lastValues   map[string]map[interface{}]*protocol.Point
+	defaultValue *protocol.FieldValue
 }
 
 func (self *DerivativeAggregator) AggregatePoint(series string, group interface{}, p *protocol.Point) error {
@@ -240,11 +271,15 @@ func (self *DerivativeAggregator) GetValues(series string, group interface{}) []
 				&protocol.FieldValue{DoubleValue: &derivative},
 			},
 		}
+	} else if self.defaultValue != nil {
+		return [][]*protocol.FieldValue{
+			[]*protocol.FieldValue{self.defaultValue},
+		}
 	}
 	return [][]*protocol.FieldValue{}
 }
 
-func NewDerivativeAggregator(q *parser.SelectQuery, v *parser.Value) (Aggregator, error) {
+func NewDerivativeAggregator(q *parser.SelectQuery, v *parser.Value, defaultValue *parser.Value) (Aggregator, error) {
 	if len(v.Elems) != 1 {
 		return nil, common.NewQueryError(common.WrongNumberOfArguments, "function derivative() requires exactly one argument")
 	}
@@ -253,12 +288,18 @@ func NewDerivativeAggregator(q *parser.SelectQuery, v *parser.Value) (Aggregator
 		return nil, common.NewQueryError(common.InvalidArgument, "function derivative() doesn't work with wildcards")
 	}
 
+	wrappedDefaultValue, err := wrapDefaultValue(defaultValue)
+	if err != nil {
+		return nil, err
+	}
+
 	return &DerivativeAggregator{
 		AbstractAggregator: AbstractAggregator{
 			value: v.Elems[0],
 		},
-		firstValues: make(map[string]map[interface{}]*protocol.Point),
-		lastValues:  make(map[string]map[interface{}]*protocol.Point),
+		firstValues:  make(map[string]map[interface{}]*protocol.Point),
+		lastValues:   make(map[string]map[interface{}]*protocol.Point),
+		defaultValue: wrappedDefaultValue,
 	}, nil
 }
 
@@ -323,7 +364,7 @@ func (self *HistogramAggregator) GetValues(series string, group interface{}) [][
 	return returnValues
 }
 
-func NewHistogramAggregator(q *parser.SelectQuery, v *parser.Value) (Aggregator, error) {
+func NewHistogramAggregator(q *parser.SelectQuery, v *parser.Value, defaultValue *parser.Value) (Aggregator, error) {
 	if len(v.Elems) < 1 {
 		return nil, common.NewQueryError(common.WrongNumberOfArguments, "function histogram() requires at least one arguments")
 	}
@@ -365,7 +406,8 @@ func NewHistogramAggregator(q *parser.SelectQuery, v *parser.Value) (Aggregator,
 //
 
 type CountAggregator struct {
-	counts map[string]map[interface{}]int32
+	defaultValue *protocol.FieldValue
+	counts       map[string]map[interface{}]int32
 }
 
 func (self *CountAggregator) AggregatePoint(series string, group interface{}, p *protocol.Point) error {
@@ -384,17 +426,22 @@ func (self *CountAggregator) ColumnNames() []string {
 
 func (self *CountAggregator) GetValues(series string, group interface{}) [][]*protocol.FieldValue {
 	returnValues := [][]*protocol.FieldValue{}
-	value := int64(self.counts[series][group])
-	returnValues = append(returnValues, []*protocol.FieldValue{
-		&protocol.FieldValue{Int64Value: &value},
-	})
+	_value, ok := self.counts[series][group]
+	value := int64(_value)
+	if !ok {
+		returnValues = append(returnValues, []*protocol.FieldValue{self.defaultValue})
+	} else {
+		returnValues = append(returnValues, []*protocol.FieldValue{
+			&protocol.FieldValue{Int64Value: &value},
+		})
+	}
 
 	return returnValues
 }
 
 func (self *CountAggregator) InitializeFieldsMetadata(series *protocol.Series) error { return nil }
 
-func NewCountAggregator(q *parser.SelectQuery, v *parser.Value) (Aggregator, error) {
+func NewCountAggregator(q *parser.SelectQuery, v *parser.Value, defaultValue *parser.Value) (Aggregator, error) {
 	if len(v.Elems) != 1 {
 		return nil, common.NewQueryError(common.WrongNumberOfArguments, "function count() requires exactly one argument")
 	}
@@ -403,20 +450,25 @@ func NewCountAggregator(q *parser.SelectQuery, v *parser.Value) (Aggregator, err
 		return nil, common.NewQueryError(common.InvalidArgument, "function count() doesn't work with wildcards")
 	}
 
+	wrappedDefaultValue, err := wrapDefaultValue(defaultValue)
+	if err != nil {
+		return nil, err
+	}
+
 	if v.Elems[0].Type != parser.ValueSimpleName {
 		innerName := v.Elems[0].Name
 		init := registeredAggregators[innerName]
 		if init == nil {
 			return nil, common.NewQueryError(common.InvalidArgument, fmt.Sprintf("Unknown function %s", innerName))
 		}
-		inner, err := init(q, v.Elems[0])
+		inner, err := init(q, v.Elems[0], defaultValue)
 		if err != nil {
 			return nil, err
 		}
-		return NewCompositeAggregator(&CountAggregator{make(map[string]map[interface{}]int32)}, inner)
+		return NewCompositeAggregator(&CountAggregator{wrappedDefaultValue, make(map[string]map[interface{}]int32)}, inner)
 	}
 
-	return &CountAggregator{make(map[string]map[interface{}]int32)}, nil
+	return &CountAggregator{wrappedDefaultValue, make(map[string]map[interface{}]int32)}, nil
 }
 
 //
@@ -483,8 +535,9 @@ func NewTimestampAggregator(query *parser.SelectQuery, _ *parser.Value) (Aggrega
 
 type MeanAggregator struct {
 	AbstractAggregator
-	means  map[string]map[interface{}]float64
-	counts map[string]map[interface{}]int
+	means        map[string]map[interface{}]float64
+	counts       map[string]map[interface{}]int
+	defaultValue *protocol.FieldValue
 }
 
 func (self *MeanAggregator) AggregatePoint(series string, group interface{}, p *protocol.Point) error {
@@ -527,31 +580,46 @@ func (self *MeanAggregator) ColumnNames() []string {
 
 func (self *MeanAggregator) GetValues(series string, group interface{}) [][]*protocol.FieldValue {
 	returnValues := [][]*protocol.FieldValue{}
-	mean := self.means[series][group]
-	returnValues = append(returnValues, []*protocol.FieldValue{
-		&protocol.FieldValue{DoubleValue: &mean},
-	})
+	mean, ok := self.means[series][group]
+	if !ok {
+		returnValues = append(returnValues, []*protocol.FieldValue{self.defaultValue})
+	} else {
+		returnValues = append(returnValues, []*protocol.FieldValue{
+			&protocol.FieldValue{DoubleValue: &mean},
+		})
+	}
 
 	return returnValues
 }
 
-func NewMeanAggregator(_ *parser.SelectQuery, value *parser.Value) (Aggregator, error) {
+func NewMeanAggregator(_ *parser.SelectQuery, value *parser.Value, defaultValue *parser.Value) (Aggregator, error) {
 	if len(value.Elems) != 1 {
 		return nil, common.NewQueryError(common.WrongNumberOfArguments, "function mean() requires exactly one argument")
+	}
+
+	wrappedDefaultValue, err := wrapDefaultValue(defaultValue)
+	if err != nil {
+		return nil, err
 	}
 
 	return &MeanAggregator{
 		AbstractAggregator: AbstractAggregator{
 			value: value.Elems[0],
 		},
-		means:  make(map[string]map[interface{}]float64),
-		counts: make(map[string]map[interface{}]int),
+		means:        make(map[string]map[interface{}]float64),
+		counts:       make(map[string]map[interface{}]int),
+		defaultValue: wrappedDefaultValue,
 	}, nil
 }
 
-func NewMedianAggregator(_ *parser.SelectQuery, value *parser.Value) (Aggregator, error) {
+func NewMedianAggregator(_ *parser.SelectQuery, value *parser.Value, defaultValue *parser.Value) (Aggregator, error) {
 	if len(value.Elems) != 1 {
 		return nil, common.NewQueryError(common.WrongNumberOfArguments, "function median() requires exactly one argument")
+	}
+
+	wrappedDefaultValue, err := wrapDefaultValue(defaultValue)
+	if err != nil {
+		return nil, err
 	}
 
 	return &PercentileAggregator{
@@ -561,6 +629,7 @@ func NewMedianAggregator(_ *parser.SelectQuery, value *parser.Value) (Aggregator
 		functionName: "median",
 		percentile:   50.0,
 		float_values: make(map[string]map[interface{}][]float64),
+		defaultValue: wrappedDefaultValue,
 	}, nil
 }
 
@@ -573,6 +642,7 @@ type PercentileAggregator struct {
 	functionName string
 	percentile   float64
 	float_values map[string]map[interface{}][]float64
+	defaultValue *protocol.FieldValue
 }
 
 func (self *PercentileAggregator) AggregatePoint(series string, group interface{}, p *protocol.Point) error {
@@ -607,11 +677,16 @@ func (self *PercentileAggregator) ColumnNames() []string {
 
 func (self *PercentileAggregator) GetValues(series string, group interface{}) [][]*protocol.FieldValue {
 	returnValues := [][]*protocol.FieldValue{}
+	values := self.float_values[series][group]
 
-	sort.Float64s(self.float_values[series][group])
-	length := len(self.float_values[series][group])
+	if len(values) == 0 {
+		returnValues = append(returnValues, []*protocol.FieldValue{self.defaultValue})
+	}
+
+	sort.Float64s(values)
+	length := len(values)
 	index := int(math.Floor(float64(length)*self.percentile/100.0+0.5)) - 1
-	point := self.float_values[series][group][index]
+	point := values[index]
 	returnValues = append(returnValues, []*protocol.FieldValue{
 		&protocol.FieldValue{DoubleValue: &point},
 	})
@@ -619,7 +694,7 @@ func (self *PercentileAggregator) GetValues(series string, group interface{}) []
 	return returnValues
 }
 
-func NewPercentileAggregator(_ *parser.SelectQuery, value *parser.Value) (Aggregator, error) {
+func NewPercentileAggregator(_ *parser.SelectQuery, value *parser.Value, defaultValue *parser.Value) (Aggregator, error) {
 	if len(value.Elems) != 2 {
 		return nil, common.NewQueryError(common.WrongNumberOfArguments, "function percentile() requires exactly two arguments")
 	}
@@ -629,6 +704,11 @@ func NewPercentileAggregator(_ *parser.SelectQuery, value *parser.Value) (Aggreg
 		return nil, common.NewQueryError(common.InvalidArgument, "function percentile() requires a numeric second argument between 0 and 100")
 	}
 
+	wrappedDefaultValue, err := wrapDefaultValue(defaultValue)
+	if err != nil {
+		return nil, err
+	}
+
 	return &PercentileAggregator{
 		AbstractAggregator: AbstractAggregator{
 			value: value.Elems[0],
@@ -636,6 +716,7 @@ func NewPercentileAggregator(_ *parser.SelectQuery, value *parser.Value) (Aggreg
 		functionName: "percentile",
 		percentile:   percentile,
 		float_values: make(map[string]map[interface{}][]float64),
+		defaultValue: wrappedDefaultValue,
 	}, nil
 }
 
@@ -645,7 +726,8 @@ func NewPercentileAggregator(_ *parser.SelectQuery, value *parser.Value) (Aggreg
 
 type ModeAggregator struct {
 	AbstractAggregator
-	counts map[string]map[interface{}]map[float64]int
+	counts       map[string]map[interface{}]map[float64]int
+	defaultValue *protocol.FieldValue
 }
 
 func (self *ModeAggregator) AggregatePoint(series string, group interface{}, p *protocol.Point) error {
@@ -711,19 +793,29 @@ func (self *ModeAggregator) GetValues(series string, group interface{}) [][]*pro
 		})
 	}
 
+	if len(values) == 0 {
+		returnValues = append(returnValues, []*protocol.FieldValue{self.defaultValue})
+	}
+
 	return returnValues
 }
 
-func NewModeAggregator(_ *parser.SelectQuery, value *parser.Value) (Aggregator, error) {
+func NewModeAggregator(_ *parser.SelectQuery, value *parser.Value, defaultValue *parser.Value) (Aggregator, error) {
 	if len(value.Elems) != 1 {
 		return nil, common.NewQueryError(common.WrongNumberOfArguments, "function mode() requires exactly one argument")
+	}
+
+	wrappedDefaultValue, err := wrapDefaultValue(defaultValue)
+	if err != nil {
+		return nil, err
 	}
 
 	return &ModeAggregator{
 		AbstractAggregator: AbstractAggregator{
 			value: value.Elems[0],
 		},
-		counts: make(map[string]map[interface{}]map[float64]int),
+		counts:       make(map[string]map[interface{}]map[float64]int),
+		defaultValue: wrappedDefaultValue,
 	}, nil
 }
 
@@ -733,7 +825,8 @@ func NewModeAggregator(_ *parser.SelectQuery, value *parser.Value) (Aggregator, 
 
 type DistinctAggregator struct {
 	AbstractAggregator
-	counts map[string]map[interface{}]map[interface{}]int
+	counts       map[string]map[interface{}]map[interface{}]int
+	defaultValue *protocol.FieldValue
 }
 
 func (self *DistinctAggregator) AggregatePoint(series string, group interface{}, p *protocol.Point) error {
@@ -781,7 +874,12 @@ func (self *DistinctAggregator) ColumnNames() []string {
 func (self *DistinctAggregator) GetValues(series string, group interface{}) [][]*protocol.FieldValue {
 	returnValues := [][]*protocol.FieldValue{}
 
-	for value, _ := range self.counts[series][group] {
+	values := self.counts[series][group]
+	if len(values) == 0 {
+		returnValues = append(returnValues, []*protocol.FieldValue{self.defaultValue})
+	}
+
+	for value, _ := range values {
 		switch v := value.(type) {
 		case int:
 			i := int64(v)
@@ -798,12 +896,18 @@ func (self *DistinctAggregator) GetValues(series string, group interface{}) [][]
 	return returnValues
 }
 
-func NewDistinctAggregator(_ *parser.SelectQuery, value *parser.Value) (Aggregator, error) {
+func NewDistinctAggregator(_ *parser.SelectQuery, value *parser.Value, defaultValue *parser.Value) (Aggregator, error) {
+	wrappedDefaultValue, err := wrapDefaultValue(defaultValue)
+	if err != nil {
+		return nil, err
+	}
+
 	return &DistinctAggregator{
 		AbstractAggregator: AbstractAggregator{
 			value: value.Elems[0],
 		},
-		counts: make(map[string]map[interface{}]map[interface{}]int),
+		counts:       make(map[string]map[interface{}]map[interface{}]int),
+		defaultValue: wrappedDefaultValue,
 	}, nil
 }
 
@@ -819,6 +923,7 @@ type CumulativeArithmeticAggregator struct {
 	values       map[string]map[interface{}]float64
 	operation    Operation
 	initialValue float64
+	defaultValue *protocol.FieldValue
 }
 
 func (self *CumulativeArithmeticAggregator) AggregatePoint(series string, group interface{}, p *protocol.Point) error {
@@ -850,9 +955,14 @@ func (self *CumulativeArithmeticAggregator) GetValues(series string, group inter
 	return returnValues
 }
 
-func NewCumulativeArithmeticAggregator(name string, value *parser.Value, initialValue float64, operation Operation) (Aggregator, error) {
+func NewCumulativeArithmeticAggregator(name string, value *parser.Value, initialValue float64, defaultValue *parser.Value, operation Operation) (Aggregator, error) {
 	if len(value.Elems) != 1 {
 		return nil, common.NewQueryError(common.WrongNumberOfArguments, "function max() requires only one argument")
+	}
+
+	wrappedDefaultValue, err := wrapDefaultValue(defaultValue)
+	if err != nil {
+		return nil, err
 	}
 
 	return &CumulativeArithmeticAggregator{
@@ -863,11 +973,12 @@ func NewCumulativeArithmeticAggregator(name string, value *parser.Value, initial
 		values:       make(map[string]map[interface{}]float64),
 		operation:    operation,
 		initialValue: initialValue,
+		defaultValue: wrappedDefaultValue,
 	}, nil
 }
 
-func NewMaxAggregator(_ *parser.SelectQuery, value *parser.Value) (Aggregator, error) {
-	return NewCumulativeArithmeticAggregator("max", value, -math.MaxFloat64, func(currentValue float64, p *protocol.FieldValue) float64 {
+func NewMaxAggregator(_ *parser.SelectQuery, value *parser.Value, defaultValue *parser.Value) (Aggregator, error) {
+	return NewCumulativeArithmeticAggregator("max", value, -math.MaxFloat64, defaultValue, func(currentValue float64, p *protocol.FieldValue) float64 {
 		if p.Int64Value != nil {
 			if fv := float64(*p.Int64Value); fv > currentValue {
 				return fv
@@ -881,8 +992,8 @@ func NewMaxAggregator(_ *parser.SelectQuery, value *parser.Value) (Aggregator, e
 	})
 }
 
-func NewMinAggregator(_ *parser.SelectQuery, value *parser.Value) (Aggregator, error) {
-	return NewCumulativeArithmeticAggregator("min", value, math.MaxFloat64, func(currentValue float64, p *protocol.FieldValue) float64 {
+func NewMinAggregator(_ *parser.SelectQuery, value *parser.Value, defaultValue *parser.Value) (Aggregator, error) {
+	return NewCumulativeArithmeticAggregator("min", value, math.MaxFloat64, defaultValue, func(currentValue float64, p *protocol.FieldValue) float64 {
 		if p.Int64Value != nil {
 			if fv := float64(*p.Int64Value); fv < currentValue {
 				return fv
@@ -896,8 +1007,8 @@ func NewMinAggregator(_ *parser.SelectQuery, value *parser.Value) (Aggregator, e
 	})
 }
 
-func NewSumAggregator(_ *parser.SelectQuery, value *parser.Value) (Aggregator, error) {
-	return NewCumulativeArithmeticAggregator("sum", value, 0, func(currentValue float64, p *protocol.FieldValue) float64 {
+func NewSumAggregator(_ *parser.SelectQuery, value *parser.Value, defaultValue *parser.Value) (Aggregator, error) {
+	return NewCumulativeArithmeticAggregator("sum", value, 0, defaultValue, func(currentValue float64, p *protocol.FieldValue) float64 {
 		var fv float64
 		if p.Int64Value != nil {
 			fv = float64(*p.Int64Value)
@@ -910,9 +1021,10 @@ func NewSumAggregator(_ *parser.SelectQuery, value *parser.Value) (Aggregator, e
 
 type FirstOrLastAggregator struct {
 	AbstractAggregator
-	name    string
-	isFirst bool
-	values  map[string]map[interface{}]*protocol.FieldValue
+	name         string
+	isFirst      bool
+	values       map[string]map[interface{}]*protocol.FieldValue
+	defaultValue *protocol.FieldValue
 }
 
 func (self *FirstOrLastAggregator) AggregatePoint(series string, group interface{}, p *protocol.Point) error {
@@ -944,25 +1056,31 @@ func (self *FirstOrLastAggregator) GetValues(series string, group interface{}) [
 	}
 }
 
-func NewFirstOrLastAggregator(name string, v *parser.Value, isFirst bool) (Aggregator, error) {
+func NewFirstOrLastAggregator(name string, v *parser.Value, isFirst bool, defaultValue *parser.Value) (Aggregator, error) {
 	if len(v.Elems) != 1 {
 		return nil, common.NewQueryError(common.WrongNumberOfArguments, "function max() requires only one argument")
+	}
+
+	wrappedDefaultValue, err := wrapDefaultValue(defaultValue)
+	if err != nil {
+		return nil, err
 	}
 
 	return &FirstOrLastAggregator{
 		AbstractAggregator: AbstractAggregator{
 			value: v.Elems[0],
 		},
-		name:    name,
-		isFirst: isFirst,
-		values:  make(map[string]map[interface{}]*protocol.FieldValue),
+		name:         name,
+		isFirst:      isFirst,
+		values:       make(map[string]map[interface{}]*protocol.FieldValue),
+		defaultValue: wrappedDefaultValue,
 	}, nil
 }
 
-func NewFirstAggregator(_ *parser.SelectQuery, value *parser.Value) (Aggregator, error) {
-	return NewFirstOrLastAggregator("first", value, true)
+func NewFirstAggregator(_ *parser.SelectQuery, value *parser.Value, defaultValue *parser.Value) (Aggregator, error) {
+	return NewFirstOrLastAggregator("first", value, true, defaultValue)
 }
 
-func NewLastAggregator(_ *parser.SelectQuery, value *parser.Value) (Aggregator, error) {
-	return NewFirstOrLastAggregator("last", value, false)
+func NewLastAggregator(_ *parser.SelectQuery, value *parser.Value, defaultValue *parser.Value) (Aggregator, error) {
+	return NewFirstOrLastAggregator("last", value, false, defaultValue)
 }