+++ CHANGELOG.md
@@ -166,7 +166,8 @@
 - [Issue #61](https://github.com/influxdb/influxdb/issues/61). Remove default limits from queries
 - [Issue #118](https://github.com/influxdb/influxdb/issues/118). Make column names starting with '_' legal
 - [Issue #121](https://github.com/influxdb/influxdb/issues/121). Don't fall back to the cluster admin auth if the db user auth fails
-- [Issue #117](https://github.com/influxdb/influxdb/issues/117). Return error on delete queries with where condition that don't have time
+- [Issue #127](https://github.com/influxdb/influxdb/issues/127). Return error on delete queries with where condition that don't have time
+- [Issue #117](https://github.com/influxdb/influxdb/issues/117). Fill empty groups with default values
 
 ### Deprecated